+++ rust/src/ark/send.rs
@@ -7,19 +7,19 @@ pub use bitcoin::Amount;
 pub use bitcoin::Txid;
 
 impl ArkClient {
-    pub async fn send_on_chain(&self, address: String, sats: u64) -> Result<Txid> {
+    pub async fn send_on_chain(&self, address: String, sats: i64) -> Result<Txid> {
         let address = Address::from_str(address.as_str())?;
         let txid = self.inner
-            .send_on_chain(address.assume_checked(), Amount::from_sat(sats))
+            .send_on_chain(address.assume_checked(), Amount::from_sat(sats as u64))
             .await
             .map_err(|e| anyhow!("Failed sending onchain {e:#}"))?;
         Ok(txid)
     }
 
-    pub async fn send_off_chain(&self, address: String, sats: u64) -> Result<Txid> {
+    pub async fn send_off_chain(&self, address: String, sats: i64) -> Result<Txid> {
         let address = ark_core::ArkAddress::decode(address.as_str())?;
         let txid = self.inner
-            .send_vtxo(address, Amount::from_sat(sats))
+            .send_vtxo(address, Amount::from_sat(sats as u64))
             .await
             .map_err(|e| anyhow!("Failed sending offchain {e:#}"))?;
         Ok(txid)
